The client and server are now able to communicate successfully for all implemented FUSE operations. Tested "cd", "ls", "touch", "rm", "mkdir" "rm -f", "rm -fr", "echo", "cat" operations on the mount point and all operations are replicated at the server directory!
1. client.py

Removed the necessity for a root dir. Each FUSE operations function now
sends the command over to the server to be executed.

2. server.py

Added functionality to map the client file/dir request to a local file/dir.
Each DFS operation now sends the result of execution of the command sent by
client back to the client.

3. common.py

Added protocol related function to enable client and server to communicate.
JSON is used to serialize and deserialize the command and request.